// src/app/Model/seguimiento.ts
export interface Seguimiento {
    ID_SEGUIMIENTOS?: number;
    EPS?: string;
    FECHA_REQUERIMIENTO?: string;
    MEDIO?: string;
    TIPO_REQUERIMIENTO?: string;
    TITULO_REQUERIMIENTO?: string;
    DESCRIPCION_REQUERIMIENTO?: string;
    AREA_VALIDACION?: string;
    ESTADO?: string
    FECHA_ENTREGA?: string;
    FECHA_FINALIZACION?: string;
    SEGUIMIENTO?: string;
    ID_REGISTRO?: string;
    ID_PERFIL?: string;
    RUTA_SOPORTES?: string;
    Categoria?: string;
    USUARIO_CREACION?:string;
    Dias_Entrega?:string;
}